generate case classes with arguments; correct false plurality
Keep using case classes for variants. When it become necessary
to use smart constructors/extractors, put all the case classes
inside object __${datatype_name}_case_classes__ and refer to them
from outside. Do NOT put the trait in there, or else the types
will be ugly. Also generate type synonyms for the variants and
refer to the synonyms in smart constructors.