generate traversable instances for algebraic functors
it requires making the argument of applicative functors covariant.
doing so makes code generator easier to implement and more general
at the same time.

code generator is easier to implement because we can generate

  applicativeInstance.pure[Nil](Nil)

instead of

  applicativeInstance.pure[ListT[Nil, Cons[_, _]]](Nil)

code generator is more general because it can also operate
on datatypes that are not sums of products.