enforce path-dependent traversals everywhere
This is done to corroborate my claim that Scala is better
than Haskell when it comes to encoding functors. Indeed,
there are fewer type annotations at use site.

TODO: use path-dependent type signatures in case studies
Compos & Scrap-your-boilerplate.

The TODO item is not urgent. It can be done after writing
the introduction to traversable functors in Scala.