standardize subcategory bound naming convention
It used to be that Traversable has `type Cat` as subcategory
bound, Traversable2 has Cat1 and Cat2, Traversable3 has Cat1
and Cat2 and Cat3, and so on. Problem is, the name of the sub-
category bound depends _both_ on the index _and_ on the arity
of the functor. Without reification, the arity of a subfunctor
cannot be known at functor composition time. Thus we switch
to a uniform naming scheme:

Traversable  has Cat0
Traversable2 has Cat0, Cat1
Traversable3 has Cat0, Cat1, Cat2
etc.