permit multiple bounds
It is necessary if we want to use a projection multiple times
in a record, say App(x, x). The deleted comment is correct only
in the case where compositeK's used to build a variant. I forgot
that it builds records too.

After this change, we delegate the task of figuring out subcat
bounds to Scala type checker. It is possible to write functors
mapping a variable to 2 or more positions in a variant (this
was impossible previously), but the type checker will catch
it upon use---the intersection of two records (distinct, final
classes) is Nothing.