embrace isorecursion
Made fixed points of data decl explicit as well, so that
declaration of datatypes mimics use of functors.

They look like this:

@data    def List [A] = Fix(list =>
           ListT { Nil ; Cons(head = A, tail = list) })

@functor def elemF[A] = Fix(list =>
           ListT { Nil ; Cons(head = A, tail = list) })

Synonyms: pending.

Preprocessor: made obsolete.

At this point, we let go of all pretentions of equirecursion.
Users are fully aware of mu-types. Since it's impossible to
hide mu-types completely, we've opted to not hide it at all,
not even to the extent Haskell hides them.