attempt to remove subcategory bounds
It succeeds to the point of generating classes.
Code/functor generation could be useful for future reference.

Coercion breaks. Horribly. Making the end-result unusable.

Subcategory bounds are not pretty, but they have the following
advantages.

1. Records are scala case classes, record fields are record
   methods.

2. They are sufficient for demonstrating the necessity
   of equirecursive types.

Seeing how difficult it is to implement coercion, we should
focus on supporting equirecursive types after this.